MAPS Location detection enabled (showing the ETA)
These changes made are in Store Comparison, both activity_maps xml, MapActivity.java, manifest file. Maps activity.java is in the maps package folder with a json parser file which is required for the map to traverse the lat and longitudes.  local file contains the google cloud enabled API Key.